day4/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
click==8.1.7
fastapi==0.114.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
joblib==1.4.2
numpy==2.1.1
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
scikit-learn==1.5.1
scipy==1.14.1
sniffio==1.3.1
starlette==0.38.5
threadpoolctl==3.5.0
typing_extensions==4.12.2
uvicorn==0.30.6
